Half Finished :)

+++ utils/ft_strlen.c
@@ -6,13 +6,13 @@
 /*   By: ehakam <ehakam@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/12/28 17:55:25 by ehakam            #+#    #+#             */
-/*   Updated: 2020/12/30 15:40:00 by ehakam           ###   ########.fr       */
+/*   Updated: 2020/12/30 16:08:51 by ehakam           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../include/cubengine.h"
 
-size_t	ft_strlen(const char *s)
+size_t		ft_strlen(const char *s)
 {
 	unsigned int	i;
 
@@ -22,7 +22,7 @@ size_t	ft_strlen(const char *s)
 	return (i);
 }
 
-int		ft_strlen_2d(t_str *str)
+size_t		ft_strlen_2d(t_str *str)
 {
 	int		i;
